Refactor rendering logic and add border support
Removed the PdfContainerRenderer class and replaced it with the PageContainerRenderer, including the addition of needed properties.  Additionally, added classes and methods needed for drawing borders on containers, including new properties for margins in the ComponentBox class and the Border class within that for describing a border's properties. Further changes involve adjustments to several other classes to accommodate these changes.